+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Added: Deploy snippets to plugin from Code Snippets Cloud app. (PRO)
 * Added: Bundles of Joy! Search and download Snippet Bundles in one go direct from Code Snippets Cloud. (PRO)
 * Fixed: Error when attempting to update network shared snippets after saving. [[#](https://wordpress.org/support/topic/activating-snippets-breaks-on-wordpress-6-3/)]
+* Improved: Redirect to snippets table when deleting snippet from the edit menu.
+* Improved: Scroll new notices into view on edit menu.
 
 ## 3.4.2 (05 Jul 2023)
 * Fixed: Issue causing export process to fail with fatal error. [[#](https://wordpress.org/support/topic/critical-error-on-exporting-snippets/)]